refactor: move base_controllers of admins to root api dir
